+++ modules/alloy/src/main/ts/ephox/alloy/api/Main.ts
@@ -109,6 +109,7 @@ import { ToolbarGroup } from './ui/ToolbarGroup';
 import { TouchMenu } from './ui/TouchMenu';
 import { Typeahead } from './ui/Typeahead';
 import * as UiSketcher from './ui/UiSketcher';
+import * as DraggingTypes from '../dragging/common/DraggingTypes';
 import {
   AnchorSpec,
   NodeAnchorSpec,
@@ -278,6 +279,7 @@ export {
   SelectionAnchorSpec,
   HotspotAnchorSpec,
   SubmenuAnchorSpec,
+  DraggingTypes,
 
   FocusInsideModes,