+++ modules/alloy/src/main/ts/ephox/alloy/dragging/common/BlockerUtils.ts
@@ -1,6 +1,8 @@
-import { Attr, Css, Traverse, Node } from '@ephox/sugar';
+import { Container } from '@ephox/alloy';
+import { Attr, Css, Node, Traverse } from '@ephox/sugar';
 
 import { AlloyComponent } from '../../api/component/ComponentApi';
+import { AlloyEventRecord } from '../../api/events/AlloyEvents';
 
 const initialAttribute = 'data-initial-z-index';
 
@@ -43,7 +45,28 @@ const discard = (blocker: AlloyComponent): void => {
   blocker.getSystem().removeFromGui(blocker);
 };
 
+const createComponent = (component: AlloyComponent, blockerClass: string, blockerEvents: AlloyEventRecord) => {
+  return component.getSystem().build(
+    Container.sketch({
+      dom: {
+        // Probably consider doing with classes?
+        styles: {
+          'left': '0px',
+          'top': '0px',
+          'width': '100%',
+          'height': '100%',
+          'position': 'fixed',
+          'z-index': '1000000000000000'
+        },
+        classes: [ blockerClass ]
+      },
+      events: blockerEvents
+    })
+  );
+};
+
 export {
+  createComponent,
   instigate,
   discard
 };