+++ modules/alloy/changelog.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+# [5.0.2] - 2019-11-11
+
+### Fixed
+- Fixed `TouchDragging` behaviour triggering drag on any document touchmove event.
+- Fixed dragging updating start state on window scroll, when dragging hadn't started.
+
 # [5.0.1] - 2019-10-25
 
 ### Fixed